Setup basic layout for hospital card

+++ EMS_Android_App/app/src/main/java/com/jia0340/ems_android_app/models/Hospital.java
@@ -10,10 +10,10 @@ public class Hospital {
     private String name;
     private NedocsScore nedocsScore;
     private HospitalType hospitalType;
-    private float distance;
+    private double distance;
     private boolean hasDiversion;
 
-    public Hospital(String name, NedocsScore nedocsScore, HospitalType hospitalType, float distance,
+    public Hospital(String name, NedocsScore nedocsScore, HospitalType hospitalType, double distance,
                     boolean hasDiversion) {
         this.name = name;
         this.nedocsScore = nedocsScore;
